# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: QA Automation Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Auth token and user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Users
...
  /audits:
    get:
      operationId: listAudits
      tags: [audits]
      summary: List audit runs
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of audit runs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditRun"
    post:
      operationId: createAudit
      tags: [audits]
      summary: Start a new audit run
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuditInput"
      responses:
        "201":
          description: Created audit run
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditRun"
  /audits/{id}:
    get:
      operationId: getAudit
      tags: [audits]
      summary: Get audit run by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Audit run
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditRun"
  /audits/{id}/cancel:
    post:
      operationId: cancelAudit
      tags: [audits]
      summary: Cancel a running audit
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cancelled audit run
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditRun"
  /audits/{id}/ai-analysis:
    get:
      operationId: getAuditAiAnalysis
      tags: [audits]
      summary: Get AI analysis for an audit (root cause, suggested fixes, risk)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: AI analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditAiAnalysis"

  # Bugs
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"

    # Users
    User:
      type: object
      required: [id, email, name, role, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, tester]
        createdAt:
          type: string
          format: date-time
...
    AuditRun:
      type: object
      required: [id, projectId, status, createdAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        projectName:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, running, completed, failed, cancelled]
        startedAt:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
        durationMs:
          type: ["integer", "null"]
        overallScore:
          type: ["number", "null"]
        bugsFound:
          type: integer
        performanceScore:
          type: ["number", "null"]
        accessibilityScore:
          type: ["number", "null"]
        seoScore:
          type: ["number", "null"]
        bestPracticesScore:
          type: ["number", "null"]
        findings:
          type: ["object", "null"]
        aiSummary:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        createdById:
          type: ["integer", "null"]
    AuditInput:
      type: object
      required: [projectId]
      properties:
        projectId:
          type: integer
    AuditAiAnalysis:
      type: object
      required: [auditRunId, summary, rootCauseAnalysis, suggestedFixes, riskAssessment]
      properties:
        auditRunId:
          type: integer
        summary:
          type: string
        rootCauseAnalysis:
          type: array
          items:
            type: object
            properties:
              issue:
                type: string
              cause:
                type: string
              impact:
                type: string
        suggestedFixes:
          type: array
          items:
            type: object
            properties:
              fix:
                type: string
              priority:
                type: string
              effort:
                type: string
        riskAssessment:
          type: object
          properties:
            level:
              type: string
              enum: [low, medium, high, critical]
            summary:
              type: string
            factors:
              type: array
              items:
                type: string

    # Bugs
...